factor out CheckableResultTable
